// src/app/pages/permisotaxiasignados/components/permisotaxiasignados-table/permisotaxiasignados.interface.ts
export interface PermisotaxiasignadosInterface {
   idpermisotaxiasignado?: number;
   status?: string;
   fecha?: string;
   chofer_idchofer?: number;
   vehiculo_idvehiculo?: number;
   permisotaxi_idpermisotaxi?: number;
   baja?: boolean;
   created_by?: number;
   created_at?: string;
   modified_at?: string;
}